app/src/main/java/com/example/feliperezende/testeteste/MainActivity.kt
package com.example.feliperezende.testeteste

import android.content.Intent
import android.support.v7.app.AppCompatActivity
import android.os.Bundle
import android.widget.Toast
import kotlinx.android.synthetic.main.activity_main.*

class MainActivity : AppCompatActivity() {

    override fun onCreate(savedInstanceState: Bundle?) {
        super.onCreate(savedInstanceState)
        setContentView(R.layout.activity_main)


        startNextActivity.setOnClickListener {
            var intent = Intent(this, SecondActivity::class.java)
            startActivity(intent)
        }

        showToast.setOnClickListener {
            Toast.makeText(this, "Lala", Toast.LENGTH_SHORT).show()
        }
    }
}
